vehicles: add avail as an argument to vehicle_part_range
also add get_avail_parts(), and make most functions that call get_parts()
call get_avail_parts() if that is more appropriate.

check that parts that should only affect the vehicle is they are on a
carried vehicle, like wheels and boat boards, aren't cached in refresh()
if they are not is_available.