vehicles: remove some redundant code
play_music and play_chimes are called in thrust() and in idle().  Since
idle() is guaranteed to be called, remove the calls in thrust().

power_parts() adds camera_epower, even though camera_epower has been
replaced with the ENABLED_DRAINS_EPOWER. The same with alarm_power and
tracking_power.  Remove all that code and use the existing flags.

Add the ENABLED_DRAINS_EPOWER to the alarm system and reduce the power
consumption to 30 watts.