weapons: add kriegsmesser and falchions as axelike swords
Add a variant of the longsword that is a little more of a chopping
blade.

Add a scaled down version of the kriegsmesser as an improved khopesh.

Also, bring the longsword up to a reference DPS of 25 by slightly
reducing bashing damage and increasing cutting damage.  Update the
to_hit value to use the new, preferred syntax.